Exercício de TiposPagamento terminado

+++ 2022-05-13/SistemaPagamento/src/TipoPagamento.java
@@ -1,4 +1,4 @@
 public interface TipoPagamento {
     public int getDiasFaturamento();
-    public float getPorcentagemFinanceiraPaga();
+    public double getPorcentagemFinanceiraPaga();
 }
